Refactor fetches
