+++ packages/client/src/App/BaseMap/Sidebar/Infobox/Infobox.tsx
@@ -5,20 +5,14 @@ import {
   type SvgIconComponent,
 } from '@mui/icons-material';
 import {Button, LinearProgress, Typography} from '@mui/material';
-import React, {memo, useCallback, useEffect, useMemo, useState} from 'react';
+import React, {memo, useCallback, useEffect, useState} from 'react';
 
-import {API_SERVER_URL, ENV_VARS} from '../../../../constants';
-import {
-  areCoordsInBounds,
-  convertDdToDmsCoords,
-  filterAndJoin,
-} from '../../../../utils';
+import {areCoordsInBounds, convertDdToDmsCoords} from '../../../../utils';
 import {useBaseMapContext} from '../../../BaseMapContext';
-import type {SearchResult} from '../SearchField';
 import type {LocationInfo} from '../Sidebar.types';
+import {useFetchLocationInfo} from '../hooks';
 import './Infobox.css';
 import type {InfoboxProps} from './Infobox.types';
-import {transformFsqAddressDetailsResponse} from './Infobox.utils';
 
 /** Renders a row showing given details about the location. */
 const InfoboxDetails = ({
@@ -40,7 +34,7 @@ const InfoboxDetails = ({
 
 /** Renders information about a location. */
 const Infobox = (props: InfoboxProps) => {
-  const {searchApi, searchResult: selectedSearchResult} = props;
+  const {searchResult: selectedSearchResult} = props;
 
   const {
     currentPos,
@@ -51,93 +45,10 @@ const Infobox = (props: InfoboxProps) => {
     setEndMarker,
   } = useBaseMapContext();
 
-  // Whether the infobox contents are currently loading.
-  const [isLoading, setIsLoading] = useState(false);
+  const {isFetching, fetchLocationInfo} = useFetchLocationInfo();
 
   // Whether the selected location is out of bounds of the map.
   const [isLocationOutOfBounds, setIsLocationOutOfBounds] = useState(false);
-
-  /** Encoded fetch URL with the param `id`. */
-  const encodedFetchLocationData = useMemo(() => {
-    let url = '';
-    let options: {} | undefined = undefined;
-
-    switch (searchApi) {
-      case 'foursquare': {
-        url = 'https://api.foursquare.com/v3/address/{id}';
-        options = {
-          headers: {
-            Authorization: ENV_VARS.fsqApiKey,
-          },
-        };
-        break;
-      }
-
-      case 'nominatim':
-      default:
-        break;
-    }
-
-    return {
-      encodedUrl: encodeURIComponent(url),
-      encodedOptions: options && encodeURIComponent(JSON.stringify(options)),
-    };
-  }, [searchApi]);
-
-  /** Fetches location information for the provided search result. */
-  const fetchLocationInfo = useCallback(
-    async (searchResult: SearchResult): Promise<LocationInfo> => {
-      setIsLoading(true);
-
-      let locationInfo: LocationInfo = {
-        label: searchResult.label,
-        description: searchResult.description,
-        address: searchResult.address ?? '',
-        latitude: searchResult.latitude ?? 0,
-        longitude: searchResult.longitude ?? 0,
-        attribution: searchResult.attribution,
-      };
-
-      switch (searchApi) {
-        case 'foursquare': {
-          if (
-            searchResult.address &&
-            searchResult.latitude !== undefined &&
-            searchResult.longitude !== undefined
-          ) {
-            break;
-          }
-          const {encodedUrl, encodedOptions} = encodedFetchLocationData;
-          const responseJson = await (
-            await fetch(
-              `${API_SERVER_URL}/fetch?` +
-                filterAndJoin(
-                  [
-                    `encodedUrl=${encodedUrl}`,
-                    encodedOptions ? `encodedOptions=${encodedOptions}` : '',
-                    `id=${encodeURIComponent(searchResult.apiId ?? '')}`,
-                  ],
-                  /* sep= */ '&'
-                )
-            )
-          ).json();
-          locationInfo = {
-            ...locationInfo,
-            ...transformFsqAddressDetailsResponse(responseJson),
-          };
-          break;
-        }
-
-        case 'nominatim':
-        default:
-          break;
-      }
-
-      setIsLoading(false);
-      return locationInfo;
-    },
-    [searchApi, encodedFetchLocationData]
-  );
 
   // The selected location.
   const [selectedLocationInfo, setSelectedLocationInfo] =
@@ -205,7 +116,7 @@ const Infobox = (props: InfoboxProps) => {
   }
   return (
     <div className="infobox">
-      {!selectedLocationInfo || isLoading ? (
+      {!selectedLocationInfo || isFetching ? (
         <Typography className="infobox-loadingContainer" color="text.disabled">
           <LinearProgress className="infobox-loading" color="inherit" />
         </Typography>
