+++ packages/client/src/App/BaseMap/Sidebar/hooks/index.ts
@@ -1,0 +1,2 @@
+export {default as useFetchLocationInfo} from './useFetchLocationInfo';
+export {default as useFetchSearchResults} from './useFetchSearchResults';
