+++ packages/client/src/App/BaseMap/Sidebar/hooks/useFetchLocationInfo/index.ts
@@ -1,0 +1,2 @@
+export {default} from './useFetchLocationInfo';
+export type {LocationResult} from './useFetchLocationInfo.types';
