+++ packages/client/src/App/BaseMap/Sidebar/hooks/useFetchSearchResults/index.ts
@@ -1,0 +1,2 @@
+export {default} from './useFetchSearchResults';
+export type {SearchResult} from './useFetchSearchResults.types';
